feat: move time step limiting calcs to util
