+++ rust/crates/util/src/lib.rs
@@ -11,6 +11,7 @@ pub mod collider_bits;
 mod consts;
 mod elastic;
 mod flat;
+mod limit_time_step;
 mod panic_to_string;
 mod particle_parameters;
 mod safe_inverse;
@@ -21,6 +22,7 @@ pub use aabb::*;
 pub use consts::*;
 pub use elastic::*;
 pub use flat::*;
+pub use limit_time_step::*;
 pub use panic_to_string::*;
 pub use particle_parameters::*;
 pub use safe_inverse::*;
